refactor: split tools_memory.rs into tools.rs and memory.rs
